refactor(map_projection_loader)!: prefix package and namespace with autoware (#8420)
* add autoware_ prefix


* add autoware_ prefix


---------
