+++ launch/tier4_map_launch/launch/map.launch.xml
@@ -59,7 +59,7 @@
       <param name="pointcloud_map_path" value="$(var pointcloud_map_path)"/>
     </node>
 
-    <include file="$(find-pkg-share map_projection_loader)/launch/map_projection_loader.launch.xml">
+    <include file="$(find-pkg-share autoware_map_projection_loader)/launch/map_projection_loader.launch.xml">
       <arg name="param_path" value="$(var map_projection_loader_param_path)"/>
       <arg name="map_projector_info_path" value="$(var map_projector_info_path)"/>
       <arg name="lanelet2_map_path" value="$(var lanelet2_map_path)"/>
