+++ launch/tier4_map_launch/package.xml
@@ -18,9 +18,9 @@
   <buildtool_depend>ament_cmake_auto</buildtool_depend>
   <buildtool_depend>autoware_cmake</buildtool_depend>
 
+  <exec_depend>autoware_map_projection_loader</exec_depend>
   <exec_depend>autoware_map_tf_generator</exec_depend>
   <exec_depend>map_loader</exec_depend>
-  <exec_depend>map_projection_loader</exec_depend>
 
   <test_depend>ament_lint_auto</test_depend>
   <test_depend>autoware_lint_common</test_depend>
