+++ map/autoware_map_projection_loader/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.14)
-project(map_projection_loader)
+project(autoware_map_projection_loader)
 
 find_package(autoware_cmake REQUIRED)
 autoware_package()
@@ -12,7 +12,7 @@ ament_auto_add_library(${PROJECT_NAME} SHARED
 )
 
 rclcpp_components_register_node(${PROJECT_NAME}
-  PLUGIN "MapProjectionLoader"
+  PLUGIN "autoware::map_projection_loader::MapProjectionLoader"
   EXECUTABLE ${PROJECT_NAME}_node
   EXECUTOR SingleThreadedExecutor
 )
