+++ map/autoware_map_projection_loader/launch/map_projection_loader.launch.xml
@@ -1,10 +1,10 @@
 <launch>
-  <arg name="param_path" default="$(find-pkg-share map_projection_loader)/config/map_projection_loader.param.yaml"/>
+  <arg name="param_path" default="$(find-pkg-share autoware_map_projection_loader)/config/map_projection_loader.param.yaml"/>
 
   <arg name="map_projector_info_path" description="Path to the yaml file"/>
   <arg name="lanelet2_map_path" description="Path to the lanelet2 map file"/>
 
-  <node pkg="map_projection_loader" exec="map_projection_loader_node" output="both">
+  <node pkg="autoware_map_projection_loader" exec="autoware_map_projection_loader_node" output="both">
     <param from="$(var param_path)" allow_substs="true"/>
   </node>
 </launch>
