+++ map/autoware_map_projection_loader/schema/map_projection_loader.schema.json
