+++ map/autoware_map_projection_loader/test/data/map_projector_info_local_cartesian_utm.yaml
