+++ map/autoware_map_projection_loader/test/data/map_projector_info_transverse_mercator.yaml
